fix(publisher): error handling invalid error string (#714)
* fix: mishandled error strings throwing

* fix: do not rely on queue error to have context

* Create purple-brooms-shake.md

+++ packages/publish/src/publish.ts
@@ -50,39 +50,6 @@ export const writeUpdate = async (
 	await fs.writeFile(pkgPath, stringify(pkgJson));
 };
 
-interface PublishObject extends BumpObject {
-	error?: unknown;
-}
-
-const packPublish = async (pkg: BumpObject): Promise<PublishObject> => {
-	const npmVersion = `${pkg.name}@${pkg.bumpVersion}`;
-	const publishFlags = ['--access', 'public', '--tag', 'latest'];
-	try {
-		await writeUpdate(pkg, { version: true });
-		await execa('npm', ['publish', ...publishFlags], {
-			cwd: pkg.path,
-		});
-		await writeUpdate(pkg, { hash: true });
-	} catch (error) {
-		consola.error(`Failed to publish ${npmVersion}!`);
-		// If we get rate limited, throw the error to stop the queue
-		// The retry delay should mean all other active publishes will also fail safely
-		// @ts-ignore - This is a custom error thrown by execa
-		if (error.message.contains('429 Too Many Requests')) {
-			throw error;
-		}
-
-		// Otherwise, just reject the promise and store the error message so we can print it later
-		return Promise.reject({
-			...pkg,
-			error,
-		});
-	}
-
-	consola.success(`Successfully published ${colors.green(npmVersion)}!`);
-	return Promise.resolve(pkg);
-};
-
 interface DoGitOptions {
 	name: string;
 	config: Context;
@@ -90,12 +57,17 @@ interface DoGitOptions {
 	npmrc: string;
 }
 
+let hasRun = false;
+
 const doGitOperations = async ({
 	name,
 	config,
 	bumped,
 	npmrc,
 }: DoGitOptions) => {
+	if (hasRun) return;
+	hasRun = true;
+
 	// Cleanup .npmrc
 	await fs.remove(npmrc);
 
@@ -111,6 +83,42 @@ const doGitOperations = async ({
 	await gitPush();
 };
 
+interface PublishObject extends BumpObject {
+	error?: unknown;
+}
+
+const packPublish = async (
+	pkg: BumpObject,
+	gitOpts: DoGitOptions
+): Promise<PublishObject> => {
+	const npmVersion = `${pkg.name}@${pkg.bumpVersion}`;
+	const publishFlags = ['--access', 'public', '--tag', 'latest'];
+	try {
+		await writeUpdate(pkg, { version: true });
+		await execa('npm', ['publish', ...publishFlags], {
+			cwd: pkg.path,
+		});
+		await writeUpdate(pkg, { hash: true });
+	} catch (error) {
+		consola.error(`Failed to publish ${npmVersion}!`);
+		// @ts-ignore - This is a custom error thrown by execa
+		const errorString = error.stderr as string;
+
+		// If we get rate limited, throw the error to stop the queue
+		// The retry delay should mean all other active publishes will also fail safely
+		if (errorString.includes('429 Too Many Requests')) {
+			await doGitOperations(gitOpts);
+			throw error;
+		}
+
+		// Otherwise, just reject the promise and store the error message so we can print it later
+		return Promise.reject(errorString);
+	}
+
+	consola.success(`Successfully published ${colors.green(npmVersion)}!`);
+	return Promise.resolve(pkg);
+};
+
 export const publishPackages = async (
 	version: string,
 	options: PublishFlags
@@ -120,8 +128,6 @@ export const publishPackages = async (
 			options.forcePublish ? colors.red(colors.bold('[FORCE]')) : ''
 		}`
 	);
-
-	const queue = new PQueue({ concurrency: 8 });
 
 	// Check for required environment variables
 	await checkEnv();
@@ -144,26 +150,19 @@ export const publishPackages = async (
 		`//registry.npmjs.org/:_authToken=${process.env.NPM_TOKEN}`
 	);
 
+	const queue = new PQueue({ concurrency: 8 });
+
 	// Collect errored out packages
 	const publishArr = [];
 
 	for (const pkg of bumped) {
 		if (pkg && !pkg.noPublish) {
-			const newPkg = queue.add(() => packPublish(pkg));
+			const newPkg = queue.add(() =>
+				packPublish(pkg, { name, config, bumped, npmrc })
+			);
 			publishArr.push(newPkg);
 		}
 	}
-
-	queue.on('error', async (error) => {
-		if (error.message.includes('429 Too Many Requests')) {
-			// We still want to commit the font changes and already published packages publish hashes
-			await doGitOperations({ name, config, bumped, npmrc });
-			throw error;
-		} else {
-			consola.error("This shouldn't happen.");
-			throw error;
-		}
-	});
 
 	// We only want the build to crash on a 429 error
 	// Otherwise, we want to continue publishing, commit and print the errors at the end
@@ -175,12 +174,7 @@ export const publishPackages = async (
 		consola.error('Failed to publish the following packages:');
 		for (const error of errors) {
 			if (error.status !== 'rejected') continue; // This should never happen, but ts is complaining
-			const pkg = error.reason as PublishObject;
-			consola.error(
-				`${colors.bold(colors.red(pkg.name))} - ${colors.red(
-					(pkg.error as Error).message
-				)}`
-			);
+			consola.error(`${colors.red(error.reason)}`);
 		}
 	}
